+++ example/lib/songs.dart
@@ -33,25 +33,20 @@ class _SongsState extends State<Songs> {
                               FlatButton(
                                 onPressed: () async {
                                   List<String> songs = [];
-                                  for (int i = 0;
-                                      i < widget.songs.length;
-                                      i++) {
+                                  for (int i = 0; i < widget.songs.length; i++) {
                                     songs.add(widget.songs[i].iOSSongID);
                                   }
                                   Playify myplayer = Playify();
-                                  var res = await myplayer.setQueue(
-                                      songIDs: songs, startIndex: index);
+                                  var res = await myplayer.setQueue(songIDs: songs, startIndex: index);
                                 },
                                 color: Colors.blueGrey[50],
-                                shape: RoundedRectangleBorder(
-                                    borderRadius: BorderRadius.circular(20)),
+                                shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(20)),
                                 child: Padding(
                                   padding: EdgeInsets.all(12),
                                   child: Column(
                                     children: <Widget>[
                                       Text(
-                                        widget.songs[index].trackNumber
-                                                .toString() +
+                                        widget.songs[index].trackNumber.toString() +
                                             ". " +
                                             widget.songs[index].title,
                                         style: TextStyle(